+++ src/vue/forms/PhoneNumberForm.vue
@@ -39,7 +39,7 @@
         v-ripple
         type="submit"
         class="phone-number-form__btn app__button-raised"
-        :disabled="formMixin.isDisabled || isUserPhoneNumber"
+        :disabled="formMixin.isDisabled || isPhoneNumberChanged"
       >
         <template v-if="isPhoneEnabled">
           {{ 'phone-number-form.change-btn' | globalize }}
@@ -65,7 +65,7 @@ import { ErrorHandler } from '@/js/helpers/error-handler'
 import { api, factorsManager } from '@/api'
 import { errors } from '@tokend/js-sdk'
 import { vuexTypes } from '@/vuex'
-import { mapGetters } from 'vuex'
+import { mapGetters, mapActions } from 'vuex'
 import IdentityGetter from '@/vue/mixins/identity-getter'
 import { Bus } from '@/js/helpers/event-bus'
 
@@ -97,7 +97,7 @@ export default {
       accountId: vuexTypes.accountId,
       isPhoneEnabled: vuexTypes.isPhoneEnabled,
     }),
-    isUserPhoneNumber () {
+    isPhoneNumberChanged () {
       return this.userPhoneNumer === this.form.phoneNumber
     },
   },
@@ -109,6 +109,10 @@ export default {
   },
 
   methods: {
+    ...mapActions({
+      loadFactors: vuexTypes.LOAD_FACTORS,
+      loadNewIdentities: vuexTypes.LOAD_IDENTITIES_BY_ACCOUNT_ID,
+    }),
     async submit () {
       const validationRule = this.isShowSmsCode
         ? this.isFormValid()
@@ -152,6 +156,7 @@ export default {
       try {
         await factorsManager.verifyTotpFactor(error, this.form.code)
         await this.changePhoneNumber()
+        await this.loadNewIdentities()
         this.$emit(EVENTS.submitted)
         if (this.isPhoneEnabled) {
           Bus.success('phone-number-form.phone-number-changed-msg')