+++ src/vue/pages/security-page-module.js
@@ -4,6 +4,7 @@ import { ShowSeedPseudoModule } from '@/modules-arch/pseudo-modules/show-seed-ps
 import { ChangePasswordPseudoModule } from '@/modules-arch/pseudo-modules/change-password-pseudo-module'
 import { ShowNetworkPassphrasePseudoModule } from '@/modules-arch/pseudo-modules/show-network-passphrase-pseudo-module'
 import { PhoneNumberFormPseudoModule } from '@/modules-arch/pseudo-modules/phone-number-form-pseudo-module'
+import { TelegramFormPseudoModule } from '@/modules-arch/pseudo-modules/telegram-form-pseudo-module'
 import { DefaultQuoteAssetPseudoModule } from '@/modules-arch/pseudo-modules/default-quote-asset-pseudo-module'
 
 export class SecurityPageModule extends PageModuleDescriptor {
@@ -17,6 +18,7 @@ export class SecurityPageModule extends PageModuleDescriptor {
         ChangePasswordPseudoModule,
         ShowNetworkPassphrasePseudoModule,
         PhoneNumberFormPseudoModule,
+        TelegramFormPseudoModule,
         DefaultQuoteAssetPseudoModule,
       ],
     })